refactor: change controls directory
