+++ lib/src/controls/flutter_deck_controls.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_deck/src/controls/actions/actions.dart';
 import 'package:flutter_deck/src/flutter_deck.dart';
 import 'package:flutter_deck/src/flutter_deck_layout.dart';
 import 'package:flutter_deck/src/theme/flutter_deck_theme.dart';
-import 'package:flutter_deck/src/widgets/internal/controls/actions/actions.dart';
 
 /// A widget that allows the user to control the slide deck.
 ///
